+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Next
+- Add `shake` method in `Camera`
+
 # 1.4.8
 - Improvements in process Tile in `TiledMap`.